pyproject.toml
[project]
name = "gpt-mcp"
version = "0.1.0"
description = "MCP server with advanced search and reasoning for all OpenAI models, inspired by o3-search-mcp"
readme = "README.md"
authors = [
    { name = "orangekame3", email = "miya.org.0309@gmail.com" }
]
requires-python = ">=3.8"
dependencies = [
    "mcp[cli]>=1.12.4",
    "openai>=1.99.4",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]

[project.scripts]
gpt-mcp = "src.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
